refactor: domain으로 AddressValidator 이동
