+++ android/app/src/main/java/com/woowacourse/ody/domain/validator/AddressValidator.kt
@@ -1,4 +1,4 @@
-package com.woowacourse.ody.presentation.address
+package com.woowacourse.ody.domain.validator
 
 object AddressValidator {
     private val CAPITAL_REGIONS = listOf("서울", "경기", "인천")
