+++ android/app/src/main/java/com/woowacourse/ody/presentation/meetinginfo/MeetingInfoViewModel.kt
@@ -6,7 +6,7 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.map
 import com.woowacourse.ody.domain.model.GeoLocation
-import com.woowacourse.ody.presentation.address.AddressValidator
+import com.woowacourse.ody.domain.validator.AddressValidator
 import com.woowacourse.ody.presentation.common.Event
 import com.woowacourse.ody.presentation.common.emit
 import java.time.LocalDateTime
